Created TagPolicy to authorize users against Tag

+++ resources/views/admin/tags/index.blade.php
@@ -1,8 +1,11 @@
 @extends('layouts.admin', ['currentPage' => 'tags'])
 
 @section('content')
-    <a class="btn btn-primary" href="{{ route('admin.tags.create') }}">Create Tag</a>
-    <table class="table mt-3">
+    @can('create', App\Models\Tag::class)
+        <a class="btn btn-primary mb-3" href="{{ route('admin.tags.create') }}">Create Tag</a>
+    @endcan
+
+    <table class="table">
         <thead>
             <tr>
                 <th scope="col">ID</th>
@@ -16,23 +19,29 @@
                     <th scope="row">{{ $tag->id }}</th>
                     <td>{{ $tag->name }}</td>
                     <td class="d-flex flex-row">
-                        <a href="{{ route('admin.tags.show', $tag) }}" class="btn btn-primary">Show Articles</a>
+                        @can('view', $tag)
+                            <a href="{{ route('admin.tags.show', $tag) }}" class="btn btn-primary">Show Articles</a>
+                        @endcan
 
-                        {{ Form::open([
-                            'method' => 'GET',
-                            'route' => ['admin.tags.edit', $tag],
-                            'class' => 'mx-1'
-                            ]) }}
-                        {{ Form::submit('Edit', ['class' => 'btn btn-warning']) }}
-                        {{ Form::close() }}
-
-                        {{ Form::open([
-                                'method' => 'DELETE',
-                                'route' => ['admin.tags.destroy', $tag],
+                        @can('update', $tag)
+                            {{ Form::open([
+                                'method' => 'GET',
+                                'route' => ['admin.tags.edit', $tag],
                                 'class' => 'mx-1'
-                            ]) }}
-                        {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
-                        {{ Form::close() }}
+                                ]) }}
+                            {{ Form::submit('Edit', ['class' => 'btn btn-warning']) }}
+                            {{ Form::close() }}
+                        @endcan
+
+                        @can('delete', $tag)
+                            {{ Form::open([
+                                    'method' => 'DELETE',
+                                    'route' => ['admin.tags.destroy', $tag],
+                                    'class' => 'mx-1'
+                                ]) }}
+                            {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
+                            {{ Form::close() }}
+                        @endcan
                     </td>
                 </tr>
             @endforeach
